fix cannot read property receiver of null

+++ server/resolvers/likeResolver.js
@@ -1,5 +1,3 @@
-import { default as Notification} from './notiResolver';
-
 const Mutation = {
   /**
    * check user authentication and create a like
@@ -46,9 +44,14 @@ const Mutation = {
     await User.findOneAndUpdate({ _id: like.user }, { $pull: { likes: like._id } });
     await Post.findOneAndUpdate({ _id: like.post }, { $pull: { likes: like.id } });
 
-    const notification = await Notification.findOneAndDelete({like: likeId});
-    // remove from user collection
-    await User.findOneAndUpdate({_id: notification.receiver}, {$pull: { notifications: notification._id}});
+        // remove from user collection
+    // we only perform the notification removal if the user unliking (lke.user)
+    // is not the post author (post.author)
+    const post = await Post.findOne({_id: like.post});
+    if(post.author != like.user){
+      const notification = await Notification.findOneAndDelete({like: likeId});
+      await User.findOneAndUpdate({_id: notification.receiver}, {$pull: { notifications: notification._id}});
+    }
 
     return like;
   },
